+++ werkzeug/contrib/__init__.py
@@ -14,3 +14,11 @@
     :copyright: (c) 2014 by the Werkzeug Team, see AUTHORS for more details.
     :license: BSD, see LICENSE for more details.
 """
+import warnings
+
+
+class WerkzeugContribDeprecationWarning(DeprecationWarning):
+    pass
+
+
+warnings.simplefilter("module", WerkzeugContribDeprecationWarning)